Implement sort options for the server and give them default options
