+++ vite-express-project/src/client/components/EntityList/EntityListContext/useSortOptions.js
@@ -1,0 +1,11 @@
+import { useState, useCallback } from 'react';
+
+export const useSortOptions = () => {
+  const [sortOptions, setSortOptions] = useState({});
+
+  return {
+    sortAttribute: sortOptions.attribute || 'name',
+    sortDirection: sortOptions.direction || 'asc',
+    setSortOptions: (attribute, direction) => setSortOptions({ attribute, direction })
+  };
+}
