+++ vite-express-project/src/client/components/EntityList/Header/SortOptions.jsx
@@ -1,0 +1,3 @@
+export const SortOptions = () => {
+  return <>Sorting/filtering options will be here</>;
+};
